Hice muchos cambios
Se imprimen bien las ventas
Los tickets tambien, con los precios de los que estan en oferta
Tambien se pueden hacer ofertas
Los codigos de barra se pueden generar
En la parte de ingresar tambien hubo cambios
Y en la parte de las ventas
Voy a seguir con la perte de empleados

+++ database/migrations/2019_10_24_003425_create_article_sale_table.php
@@ -19,8 +19,8 @@ class CreateArticleSaleTable extends Migration
             $table->integer('article_id');
             $table->integer('sale_id');
             $table->integer('amount');
-            $table->integer('cost')->nullable();
-            $table->integer('price')->nullable();
+            $table->decimal('cost')->nullable();
+            $table->decimal('price')->nullable();
 
             $table->timestamps();
         });
